fix(lsm-read): rq_bits is u8

+++ rust/lance/benches/mem_wal/fts/mem_wal_lsm_read_bench.rs
@@ -122,7 +122,7 @@ struct Args {
     rescore_factor: u32,
     vector_dim: usize,
     ivf_partitions: usize,
-    rq_bits: u32,
+    rq_bits: u8,
     nprobes: usize,
     cache_dir: PathBuf,
     output: Option<PathBuf>,
